* Fixed AMD64 compilation problems * const-corrected some function calls
git-svn-id: svn://svn.berlios.de/qlix/trunk@130 97db03ad-8736-0410-9762-8805f5232b6b

+++ widgets/QMtpDevice.cpp
@@ -239,7 +239,7 @@ void QMtpDevice::findAndRetrieveDeviceIcon()
   MTP::Folder* rootFolder = _device->RootFolder();
   count_t fileCount = rootFolder->FileCount();
 
-  QString iconPath = QString("/tmp/QlixDeviceIcon-%1").arg( (int) this);
+  QString iconPath = QString("/tmp/QlixDeviceIcon-%1").arg( (uint64_t) this);
   MTP::File* iconFile= NULL;
   for (count_t i = 0; i < fileCount; i++)
   {
